# openapi.yaml
openapi: 3.0.0
info:
  title: Ridesharing API
  description: This is main API for our small ridesharing project
  contact:
    email: 1alekseik1@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: ref
servers:
- url: /
security:
- cookieAuth: []
...
paths:
...
  /join_ride:
    post:
      tags:
      - user
      summary: Присоединиться к поездке по ID поездки
      description: Нужна авторизация. Присоединяет пользователя к поездке. ПОКА НЕ
        НУЖНО ПОДТВЕРЖДЕНИЕ СО СТОРОНЫ ВОДИТЕЛЯ
      requestBody:
        $ref: '#/components/requestBodies/joinRide'
      responses:
        200:
          description: Успех. Теперь пользователь числится в пассажирах поездки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rideID'
              examples:
                main:
                  $ref: '#/components/examples/rideID'
        400:
          description: Если каких-то данных не хватает, либо поездки с таким ID не
            существует, либо пользователь уже в поездке, либо пользователь является
            водителем этой поездки, либо поездка протухла
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                missing one parameter:
                  $ref: '#/components/examples/errorMissingOneParameter'
                missing two parameters:
                  $ref: '#/components/examples/errorMissingTwoParameters'
                no such ride:
                  $ref: '#/components/examples/errorInvalidRide'
                already in ride:
                  $ref: '#/components/examples/errorAlreadyInRide'
                is ride host:
                  $ref: '#/components/examples/errorIsRideHost'
                ride is finished:
                  $ref: '#/components/examples/errorRideIsFinished'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
      - cookieAuth: []
...
  /find_best_rides:
    post:
      tags:
      - user
      summary: Найти лучшую поездку на основе GPS
      description: Нужно обязательно передать GPS. _Да меня вообще кто-то читает?_
        Вернет список вида `({Ride}, distance)`, где `Ride` - это в объект поездки
        в формате __JSON__
      requestBody:
        $ref: '#/components/requestBodies/findBestRides'
      responses:
        200:
          description: Все ок, возвращаем отсортированный список поездок
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rideAndDistance'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
      - cookieAuth: []
...
  /get_all_rides:
    get:
      tags:
        - user
      summary: Список ВСЕХ доступных поездок
      description: Получить список всех доступных поездок.
      responses:
        200:
          description: Проблем нет, возвращаем список поездок.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/rideInfo'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
        - cookieAuth: []
  /get_my_finished_rides:
    get:
      tags:
      - user
      summary: Личный список законченных поездок
      description: Все законченные поездки, которые либо ты захостил, либо ты в них участвовал
      responses:
        200:
          description: Проблем нет, возвращаем список поездок.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/rideInfo'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
      - cookieAuth: []
  /create_ride:
    post:
      tags:
      - driver
      summary: Создать поездку
      description: Создает поездку, нужно быть водителем.
      requestBody:
        $ref: '#/components/requestBodies/createRide'
      responses:
        200:
          description: успех, поездка создана.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rideID'
              examples:
                main:
                  $ref: '#/components/examples/rideID'
        400:
          description: Каких-то данных не хватает
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_400'
              examples:
                missing one parameter:
                  $ref: '#/components/examples/errorMissingOneParameter'
                missing two parameters:
                  $ref: '#/components/examples/errorMissingTwoParameters'
        401:
          description: Нужна авторизация, либо пользователь пытается создать поездку
            не от своего имени, либо пользователь не является водителем
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                authorization required:
                  $ref: '#/components/examples/errorAuthorizationRequred'
                is not a driver:
                  $ref: '#/components/examples/errorIsNotDriver'
                host_id mismatch:
                  $ref: '#/components/examples/errorNoPermissionForUser'
      security:
      - cookieAuth: []
...
  /finish_ride:
    post:
      tags:
        - driver
      summary: Помечает поездку как завершенную
      security:
        - cookieAuth: []
      description: Должно вызываться при авторизованном водителе
      requestBody:
        $ref: '#/components/requestBodies/joinRide'
      responses:
        200:
          description: Успех, отмечена как завершена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rideID'
        400:
          description: Любая ошибка во входных данных, либо поездка еще не началась
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                invalid ride ID:
                  $ref: '#/components/examples/errorInvalidRide'
                ride hasnot started:
                  $ref: '#/components/examples/errorRideNotStarted'
        401:
          description: Пользователь не залогинен. БАГ - пока можно быть не залогиненым как водитель
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                not logged in:
                  $ref: '#/components/examples/errorAuthorizationRequred'
  /leave_ride:
    post:
      tags:
        - user
      summary: Убирает себя (текущего пользователя) из поездки.
      security:
        - cookieAuth: []
      description: Ты должен быть авторизован как пользователь.
      requestBody:
        $ref: '#/components/requestBodies/joinRide'
      responses:
        200:
          description: Успех, пользователь больше не в поездке, а она снова доступна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/rideID'
        400:
          description: Любая ошибка во входных данных, либо текущий пользователь не состоит в поездке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                invalid ride ID:
                  $ref: '#/components/examples/errorInvalidRide'
                user not in ride:
                  $ref: '#/components/examples/errorUserNotInRide'
        401:
          description: Пользователь не залогинен. БАГ - пока можно быть не залогиненым как водитель
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                not logged in:
                  $ref: '#/components/examples/errorAuthorizationRequred'
...
  /get_ride_info:
    get:
      tags:
      - user
      summary: Информация об одной поездке
      parameters:
      - name: ride_id
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        200:
          description: Проблем нет, возвращаем инфу о поездке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        400:
          description: Нет такой поездки
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                invalid ride ID:
                  $ref: '#/components/examples/errorInvalidRide'
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
      - cookieAuth: []
  /get_my_rides:
    get:
      tags:
      - user
      summary: Список всех личных незавершенных поездок
      description: Список поездок, которые ты либо хостил, либо в них участвовал
      responses:
        200:
          description: Все ок, список готов
          content:
            application/json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/rideInfo'
                  properties:
                    is_mine:
                      type: boolean
                      description: Являюсь ли я хостом этой поездки
        401:
          description: Нужна авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorBase'
              examples:
                main:
                  $ref: '#/components/examples/errorAuthorizationRequred'
      security:
      - cookieAuth: []
...
components:
  schemas:
    rideAndDistance:
      type: array
      items:
        anyOf:
        - $ref: '#/components/schemas/rideInfo'
        - $ref: '#/components/schemas/distance'
    distance:
      type: number
    errorBase:
      required:
      - name
      - value
      type: object
      properties:
        name:
          type: string
        value:
          type: string
...
    rideID:
      required:
      - ride_id
      type: object
      properties:
        ride_id:
          type: integer
    userID:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
...
    userBase:
      required:
      - email
      - first_name
      - last_name
      - phone_number
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        organizations:
          $ref: '#/components/schemas/organizationInfo'
        phone_number:
          type: string
    userExtra:
      type: object
      properties:
        photo_url:
          type: string
          description: URL в хранилище
        is_driver:
          type: boolean
          description: Является ли пользователь водителем
...
    organizationInfo:
      required:
      - latitude
      - longitude
      - name
      type: object
      properties:
        id:
          type: integer
          description: ID в БД
        name:
          type: string
          description: Название организации
        latitude:
          type: number
          description: Координата широты
        longitude:
          type: number
          description: Координата долготы
        users:
          type: array
          description: Массив ID пользователей
          items:
            type: number
        address:
          type: string
          description: Адрес организации
        description:
          type: string
          description: Описание организации
        photo_url:
          type: string
          description: Фотка организации
    rideInfo:
      required:
      - host_driver_id
      - host_driver_info
      - is_available
      - start_organzation_id
      - start_time
      - stop_latitude
      - stop_longitude
      - total_seats
      type: object
      properties:
        description:
          type: string
          description: Описание (необязательное)
        estimated_time:
          type: string
          description: Оценочное время
        host_driver_id:
          type: integer
          description: ID водителя, который начал поездку
        host_driver_info:
          allOf:
          - $ref: '#/components/schemas/userID'
          - $ref: '#/components/schemas/userBase'
          - $ref: '#/components/schemas/userExtra'
        id:
          type: integer
          description: ID поездки
        is_available:
          type: boolean
          description: Доступна или нет
        passengers:
          type: array
          description: Массив из ID всех пассажиров
          items:
            type: integer
            description: ID `User`
        start_organization:
          type: integer
          description: ID организации, откуда выезжаем
        start_organization_id:
          type: integer
          description: ID организации, откуда выезжаем
        start_time:
          type: string
          description: Время отправления
        stop_latitude:
          type: number
          description: Широта пункта назначения
        stop_longitude:
          type: number
          description: Долгота пункта назначения
        cost:
          type: string
          description: Цена поездки
        total_seats:
          type: integer
          description: Всего мест в автомобиле
        stop_address:
          type: string
          description: Адрес пункта назначения
    inline_response_400:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/errorBase'
    inline_response_200:
      required:
      - seats_available
      properties:
        seats_available:
          type: integer
...
  examples:
...
    errorIsRideHost:
      value:
        name: User is ride host with user_id
        value: 36
    errorAlreadyInRide:
      value:
        name: User is already in drive with user_id
        value: 36
    errorMissingOneParameter:
      value:
        name: Some of required fields are missing
        value:
        - first_name
    errorMissingTwoParameters:
      value:
        name: Some of required fields are missing
        value:
        - first_name
        - last_name
    rideID:
      value:
        ride_id: 36
...
    errorInvalidRide:
      value:
        name: Invalid ride with id
        value: 37
    errorRideNotStarted:
      value:
        name: Ride has not started yet
        value: <ride_id>
    errorRideIsFinished:
      value:
        name: Ride is finished
        value: 19
    errorUserNotInRide:
      value:
        name: User is not in ride with ride_id
        value: 2
...
    errorIsNotDriver:
      value:
        name: User with ID is not a driver
        value: 28
    errorAuthorizationRequred:
      value:
        name: Authorization required
        value: ""
    errorNoPermissionForUser:
      value:
        name: No permission for user
        value: 2
...
  requestBodies:
...
    findBestRides:
      content:
        application/json:
          schema:
            required:
            - destination_latitude
            - destination_longitude
            - start_date
            - start_organization_id
            - max_destination_distance_km
            type: object
            properties:
              start_date:
                type: string
                description: Дата начала. Лучше потестить фронт и бек и договориться
                  о формате
              start_organization_id:
                type: integer
              destination_latitude:
                type: number
              destination_longitude:
                type: number
              max_destination_distance_km:
                type: number
                description: Расстояние до пункта назначения, по которому отсекать
...
    createRide:
      content:
        application/json:
          schema:
            required:
            - start_organization_id
            - start_time
            - stop_latitude
            - stop_longitude
            - total_seats
            - car_id
            type: object
            properties:
              start_organization_id:
                type: integer
                description: ID организации отправления
              stop_latitude:
                type: number
                description: Широта
              stop_longitude:
                type: number
                description: Долгота
              start_time:
                type: string
                description: Время отправления
              description:
                type: string
                description: Описание поездки
              total_seats:
                type: integer
                description: Сколько мест в машине
              cost:
                type: number
                description: Сколько будет стоить
              car_id:
                type: number
                description: ID машины, на которой будем ехать
...
    joinRide:
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/rideID'
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: Session=12d12jds
      name: session
      in: cookie